Add units to PHYSICS_PARAMS for non-cosmo runs.
Currently for non-cosmo runs, amr/units.f90 always returns 1 for the
unit scales. However, we need to be able to convert to proper units for
the Iliev tests. So I've added unit parameters for non-cosmo runs to
PHYSICS_PARAMS.

Tested:
Ran a test with and without the units parameters specified in the
namelist.


git-svn-id: https://svn.physik.uzh.ch/repos/itp/ramses@139 fa3a7ada-b1ec-11df-bea4-b5d77251c905