For the radiation pressure patch, added gas heating from the momentum injected from isotropic photons.
Also changed the way multiscattering of IR photons is implemented.
Now when an IR photon is scattered (on dust), 1/1000 of the photon 
(field) energy is depleted, and the photon is scattered isotropically.
This means we subtract 1/1000 of the energy of the scattered IR photons
from the photon density, and the full energy of those same photons from
the photon flux.
 


git-svn-id: https://svn.physik.uzh.ch/repos/itp/ramses@403 fa3a7ada-b1ec-11df-bea4-b5d77251c905