Updated UV background, both homogeneous and RT-propagated, such that it is damped according to the difference in z and z_reion, just as in the ramses equilibrium cooling module.
This makes for a more gradual onset of reionisation, whereas
it happens as a very sudden jump without the damping.

Also, implemented the self-shielding threshold in the case of
the homogeneous UV background (which is now hardcoded at 1d-2 H/cc)
with a smoother function around the threshold, rather than the
previous cutoff. The same smoother function is now also used for the
density threshold for which to add the propagated UV background,
which is added below a given density.

Also one bug-fix in rt_spectra: wrong number of dimensions
in MPI-communication of propagated UV background table.