New domain_t type with data and procedures
Encapsulate data and methods related to a specific P-H domain decomposition in to a single type.

The type contains data for
  n, overload: total nr of intervals and nr of intervals per cpu rank (ncpu/ndomain)
  b: hilbert key interval tick marks (bound_key)
  r2d, d2r: mapping from cpu rank to a domain and vice versa (did not exist before)
and methods
  in_rank, get_rank: given a P-H key is it on this rank / which rank has it?
  create, destroy, copy: setup, deallocate, and copy a domain decomposition type
  pos2key, key2pos: convert from a position, level to/from a P-H key

The in_rank and get_rank methods are more readable throughout the code, and get_rank is substantially faster than the do-loop it substitutes.

Enabled permutation of P-H segments to improve connectivity between levels.